:art: split files and purify history
