+++ core/src/getFullHrefVersion/isRoute.js
@@ -1,0 +1,1 @@
+export default route => typeof route === 'object' && route.code
