+++ core/src/middleware/mapAction/index.js
@@ -1,0 +1,1 @@
+export { default } from './mapAction'
